2024 day 12 Rust solutions
Part 1 was fairly straightforward but part 2 was hard because
I couldn't come up with a good way to find the sides. Eventually
decided on simply adding all fences connected to the current
fence to a side and then ignore them in future loops. It works
but is probably not optimal.