2024 day 14 Rust solutions
This was a pretty interesting problem, part 1 was not
too hard, the movement could be easily handled with
modular arithmetic. Part 2 was a bit harder because
finding the easter egg felt a bit random. I basically
checked each position manually. Though I did notice
two pattern that showed up around every 60-100 positions
so I thought that when they overlap is when the easter
egg will show up. This made the search a bit easier and
I think it was right because eventually I could find
the Christmas tree. At least it was an interesting
problem.